feat: Add check login status logic Before show cookie consent dialog
