+++ frontend/src/app/[locale]/difficulty/page.tsx
@@ -1,59 +1,11 @@
-'use client';
+import React from 'react';
+import { AuthService } from '@/services/AuthService';
+import { CookieService } from '@/services/CookieService';
+import DifficultyClientPage from './clientPage';
 
-import React, { useEffect, useState } from 'react';
-import { useTranslations } from 'next-intl';
+export default async function DifficultyPage() {
+  const authService = new AuthService(new CookieService());
+  const isLoggedIn = await authService.isLoggedIn();
 
-import DeckListPage from '@/components/common/DeckListPage';
-import LoadingSpinner from '@/components/LoadingSpinner/LoadingSpinner';
-import { getDecks } from '@/api/decks';
-
-import { Deck } from '@/types/schemes';
-import { difficultiesInDisplayOrder } from '@/types/Category';
-import CustomDialog from '@/components/Dialogs/CustomDialog';
-
-export default function DifficultyPage() {
-  const t = useTranslations();
-  const [decks, setDecks] = useState<Deck[]>();
-  const [isCookieConsentOpen, setIsCookieConsentOpen] = useState(false);
-
-  useEffect(() => {
-    const fetchUserCards = async () => {
-      const fetchedDecks = await getDecks('difficulty');
-      if (fetchedDecks) {
-        setDecks(fetchedDecks.content);
-      }
-    };
-    fetchUserCards();
-
-    const cookieConsent = localStorage.getItem('cookieConsent');
-    if (!cookieConsent) {
-      setIsCookieConsentOpen(true);
-    }
-  }, []);
-
-  const handleCookieConsent = () => {
-    localStorage.setItem('cookieConsent', 'true');
-    setIsCookieConsentOpen(false);
-  };
-
-  if (!decks) {
-    return <LoadingSpinner />;
-  }
-
-  return (
-    <>
-      <DeckListPage
-        decks={decks}
-        categoryType="difficulty"
-        displayOrder={difficultiesInDisplayOrder}
-      />
-      <CustomDialog
-        open={isCookieConsentOpen}
-        headline={t('cookies')}
-        prompt={t('cookieConsent')}
-        firstButtonString={t('ok')}
-        firstButtonOnclick={handleCookieConsent}
-      />
-    </>
-  );
+  return <DifficultyClientPage isLoggedIn={isLoggedIn} />;
 }
